Principal & Bursar dashboards: production pass (fix + upgrade)
Audit (18-agent review, every high adversarially verified) then P0
correctness, P1 plan features, P2 polish. 54 seeded-DB runtime checks +
two zero-warning CI builds. Audit report under Principal-Bursar-Production-Pass/.

P0 — money correctness & access control (bursar):
- updateFinanceUser suspend flipped only the display mirror, never
  User.is_active — a "suspended" finance user kept full access. Now flips
  both in a transaction with a self-suspend guard (mirrors updatePrincipalUser).
- All currency FLOAT -> DECIMAL(12,2) (Fee/Payment/Expense/FeeCategory +
  Donation); db.js sets dialectOptions.decimalNumbers so DECIMAL returns as
  Number, not string. Migration: 2026-07-13-finance-money-decimal.sql.
- Receipt numbers were a bare ms timestamp (collidable) -> RCP-<school>-<ts>-
  <rand> + UNIQUE index + insert retry.
- Lost-update race on fee.amount_paid -> lock the fee row FOR UPDATE first,
  atomic increment, whole-transaction deadlock retry; guard double-rollback.
- Server-side amount validation on recordPayment/recordExpense/assignFees.
- General (fee_id null) payments now credit the student statement.
- Finance Team list filtered to finance role; management controls gated to
  school_admin/superadmin (bursar is 403 on those writes).

P1 — plan features:
- Bursar (Module 4): receipt lookup + server-side QR verify (public /verify
  gains a payment-receipt fallback); % scholarships + reasons + installment
  plans + late-fee/grace policy (overdue/late fee computed on read);
  collection rate by term/class, 12-month cash flow, budgets with
  plan-vs-actual attainment. Migration: 2026-07-13-fee-terms-and-budget.sql
  (+ FinanceBudget model).
- Principal: discipline oversight (BehaviourIncident), timetable oversight
  (TimetableSlot), teacher roster with load, school-calendar access, and
  audience-targeted announcements (all/teachers/parents/students).

P2 — UX / a11y / honesty:
- FinanceTeam mobile CSS; hardcoded amber -> theme tokens; new
  --ska-on-green/--ska-on-error so button labels stay readable in light theme;
  health score shows honest "No data yet" instead of red "Critical 0%";
  financial-anomaly alert now a real signal (24h payment > 3x 90-day avg);
  at-risk home threshold matches the At-Risk page; "AI Insights" -> "Insights".

Prod DB sync is OFF: apply both migrations by hand (already applied to dev).

+++ backend_node/src/models/Donation.js
@@ -5,7 +5,7 @@ const Donation = sequelize.define('Donation', {
   id: { type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true },
   campaign_id: { type: DataTypes.BIGINT, allowNull: false },
   donor_id: { type: DataTypes.BIGINT },
-  amount: { type: DataTypes.FLOAT },
+  amount: { type: DataTypes.DECIMAL(12, 2) },
   is_anonymous: { type: DataTypes.BOOLEAN, defaultValue: false },
   receipt_hash: { type: DataTypes.STRING(255) },
   paid_at: { type: DataTypes.DATE },
